new_mvp/requirements.txt
transformers
langchain
huggingface_hub
streamlit
faiss-cpu 
tiktoken
openai
PyPDF2
python-dotenv
kor
pymongo
selenium
webdriver-manager
llama-index             